Spent around 4 hours learning html
This is the best I've got boys I'm going to need help for the rest. I uploaded everything I've experimented with in order to get this prototype "functioning". I say "functioning" in quotes because its far from complete. Currently we have working links (although they lead to nowhere) A two body structure with a header with color and formatting based on screen size.