Wireframe of website COMPLETE
It is far from pretty but covers all testing needs for tomorrow. We can use this to verify general layout, and gain suggestions on the design (which are desperately needed). I will be going through with a few more commits in order to fix issues regarding links and images guaranteed to occur during upload.